+++ Projects/zstack/HomeAutomation/SampleSwitch/Source/zcl_samplesw.h
@@ -64,6 +64,13 @@ extern "C"
 #define SAMPLESW_ENDPOINT_RELAY4        4
 #define SAMPLESW_NUM_RELAYS             4
 
+// 4路输入状态端点 (EP 5-8, 对应 alab.switch 的 in1-in4)
+#define SAMPLESW_ENDPOINT_INPUT1        5
+#define SAMPLESW_ENDPOINT_INPUT2        6
+#define SAMPLESW_ENDPOINT_INPUT3        7
+#define SAMPLESW_ENDPOINT_INPUT4        8
+#define SAMPLESW_NUM_INPUTS             4
+
 #define LIGHT_OFF                       0x00
 #define LIGHT_ON                        0x01
 
@@ -95,9 +102,13 @@ extern SimpleDescriptionFormat_t zclSampleSw9_SimpleDesc;
 
 extern SimpleDescriptionFormat_t zclSampleSw_RelaySimpleDesc[SAMPLESW_NUM_RELAYS];
 
+extern SimpleDescriptionFormat_t zclSampleSw_InputSimpleDesc[SAMPLESW_NUM_INPUTS];
+
 extern CONST zclAttrRec_t zclSampleSw_Attrs[];
 
 extern uint8 zclSampleSw_RelayState[SAMPLESW_NUM_RELAYS];
+
+extern float zclSampleSw_InputState[SAMPLESW_NUM_INPUTS];
 
 extern uint8  zclSampleSw_OnOff;